# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Atlas API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /atlas/sources/{sourceId}/analyze:
    post:
      operationId: analyzeAtlasSource
      tags: [Atlas]
      summary: Request evidence-linked twin profile suggestions from an authorized source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceAnalysisInput"
      responses:
        "200":
          description: Suggestions for the person to review; none are applied automatically
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceAnalysis"
        "403":
          description: The source owner has not consented to profile analysis
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Storage]
      summary: Request a private presigned URL for a user-provided source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: Read a private source object after workspace authorization
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
components:
  schemas:
...
    TwinFieldKey:
      type: string
      enum:
        - role
        - expertise
        - priorities
        - communicationStyle
        - decisionMethodology
        - riskTolerance
        - delegationRules
        - approvalLimits
        - meetingBehavior
        - stakeholderRelationships
...
    SourceAnalysisInput:
      type: object
      required: [confirmAnalysis]
      properties:
        confirmAnalysis:
          type: boolean
          const: true
    ProfileSuggestion:
      type: object
      required: [key, suggestedValue, evidence, confidence]
      properties:
        key:
          $ref: "#/components/schemas/TwinFieldKey"
        suggestedValue: { type: string }
        evidence: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
    SourceAnalysis:
      type: object
      required: [sourceId, suggestions]
      properties:
        sourceId: { type: string }
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/ProfileSuggestion"
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }